Splitting out getData and getOptions from list.js
